feat: create hook for transactions
